+++ src/main/java/com/clubsis/repository/bungalow/ReservaBungalowRepository.java
@@ -7,6 +7,6 @@ import com.clubsis.model.bungalow.ReservaBungalow;
  * Created by Juan Tenorio on 1/5/2016.
  */
 
-public interface ReservaBungalowRepository extends JpaRepository<ReservaBungalow, Long> {
+public interface ReservaBungalowRepository extends JpaRepository<ReservaBungalow, Integer> {
 
 }
