+++ src/main/java/com/clubsis/repository/bungalow/TipoBungalowRepository.java
@@ -7,6 +7,6 @@ import com.clubsis.model.bungalow.TipoBungalow;
  * Created by Juan Tenorio on 1/5/2016.
  */
 
-public interface TipoBungalowRepository extends JpaRepository<TipoBungalow, Long> {
+public interface TipoBungalowRepository extends JpaRepository<TipoBungalow, Integer> {
 
 }
