+++ src/main/java/com/clubsis/repository/club/OfertaRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 /**
  * Created by Juan Tenorio on 1/5/2016.
  */
-public interface OfertaRepository extends JpaRepository<Oferta, Long> {
+public interface OfertaRepository extends JpaRepository<Oferta, Integer> {
 }
