+++ src/main/java/com/clubsis/repository/club/SancionRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 /**
  * Created by Juan Tenorio on 1/5/2016.
  */
-public interface SancionRepository extends JpaRepository<Sancion, Long> {
+public interface SancionRepository extends JpaRepository<Sancion, Integer> {
 }
