+++ src/main/java/com/clubsis/repository/club/SorteoRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 /**
  * Created by Juan Tenorio on 1/5/2016.
  */
-public interface SorteoRepository extends JpaRepository<Sorteo, Long> {
+public interface SorteoRepository extends JpaRepository<Sorteo, Integer> {
 }
