+++ src/main/java/com/clubsis/repository/pago/PagoRepository.java
@@ -6,5 +6,5 @@ import com.clubsis.model.pago.Pago;
 /**
  * Created by Juan Tenorio on 1/5/2016.
  */
-public interface PagoRepository extends JpaRepository<Pago, Long>  {
+public interface PagoRepository extends JpaRepository<Pago, Integer>  {
 }
