+++ src/main/java/com/clubsis/repository/persona/PersonaRepository.java
@@ -5,5 +5,5 @@ import com.clubsis.model.persona.Persona;
 /**
  * Created by Juan Tenorio on 1/5/2016.
  */
-public interface PersonaRepository extends JpaRepository<Persona, Long> {
+public interface PersonaRepository extends JpaRepository<Persona, Integer> {
 }
