+++ src/main/java/com/clubsis/repository/persona/SocioRepository.java
@@ -5,5 +5,5 @@ import com.clubsis.model.persona.Socio;
 /**
  * Created by Juan Tenorio on 1/5/2016.
  */
-public interface SocioRepository extends JpaRepository<Socio, Long>{
+public interface SocioRepository extends JpaRepository<Socio, Integer>{
 }
